fix: habitat

+++ app/Http/Controllers/HabitatController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\HabitatResource;
 use App\Models\Habitat;
+use App\Models\Species;
 use Illuminate\Http\Request;
 
 class HabitatController extends Controller
@@ -37,6 +38,12 @@ class HabitatController extends Controller
             'name' => ['required', 'unique:habitats'],
         ]);
 
+        $isInUse = Species::where('habitat', $habitat->name)->first();
+
+        if ($isInUse) {
+            return errorJson('Habitat is in use and cannot be updated', 409);
+        }
+
         $habitat->update($validated);
 
         return new HabitatResource($habitat);
@@ -44,6 +51,12 @@ class HabitatController extends Controller
 
     public function destroy(Habitat $habitat)
     {
+        $isInUse = Species::where('habitat', $habitat->name)->first();
+
+        if ($isInUse) {
+            return errorJson('Habitat is in use and cannot be deleted', 409);
+        }
+
         $habitat->delete();
 
         return response()->json();
